requirements.txt
tensorflow>=2.12
numpy
scikit-learn
pillow
matplotlib
streamlit
pandas